testng versions bumped to 7.0.0

+++ simple-module-example/application/src/test/java/itx/examples/java/eleven/application/tests/TestServices.java
@@ -48,7 +48,7 @@ public class TestServices {
         Assert.assertNotNull(upcommingResult);
         Float result = upcommingResult.get();
         Assert.assertNotNull(result);
-        Assert.assertEquals(result, 10F);
+        Assert.assertEquals(result, Float.valueOf(10F));
     }
 
     @Test
